🚚 rename(activity folder) and 🐛 fix(TypeScript Errors)
