+++ WEB/src/components/activity/ChooseActivities.vue
